refactor(core): move grouping policy to integrations and simplify asset processing
