+++ packages/core/src/services/assets/asset-processing-service/grouped-content-bundles.ts
@@ -1,87 +1,5 @@
-import { rapidhash } from '../../../utils/hash.ts';
-import type { AssetDefinition, ContentScriptAsset, ProcessedAsset } from './assets.types.ts';
-
-/** Shared grouped-build id prefix for page-owned content scripts. */
-export const PAGE_OWNED_CONTENT_SCRIPT_BUNDLE_PREFIX = 'ecopages';
-
-/** Returns the stable grouped-build id for one integration's page-owned content scripts. */
-export function createPageOwnedContentScriptBundleId(integrationName: string): string {
-	return `${PAGE_OWNED_CONTENT_SCRIPT_BUNDLE_PREFIX}-${integrationName}-page-content-scripts`;
-}
-
-/** Returns whether page-owned content scripts should share one grouped build for this integration. */
-export function shouldAssignPageOwnedContentScriptGroups(integrationName: string): boolean {
-	return integrationName === 'ecopages-jsx';
-}
-
-function isPageOwnedGroupableContentScript(dep: AssetDefinition): dep is ContentScriptAsset {
-	if (dep.kind !== 'script' || dep.source !== 'content' || dep.inline) {
-		return false;
-	}
-
-	if (dep.excludeFromHtml) {
-		return false;
-	}
-
-	if (dep.groupedBundle?.id) {
-		return false;
-	}
-
-	if (dep.packageRole === 'keep-separate' || dep.packageRole === 'runtime' || dep.packageRole === 'dynamic-chunk') {
-		return false;
-	}
-
-	return true;
-}
-
-function createPageOwnedGroupedEntryName(dep: ContentScriptAsset, usedEntryNames: Set<string>): string {
-	const baseName = dep.name ?? `script-${rapidhash(dep.content).toString(16)}`;
-	if (!usedEntryNames.has(baseName)) {
-		usedEntryNames.add(baseName);
-		return baseName;
-	}
-
-	let suffix = 2;
-	let candidate = `${baseName}-${suffix}`;
-	while (usedEntryNames.has(candidate)) {
-		suffix += 1;
-		candidate = `${baseName}-${suffix}`;
-	}
-
-	usedEntryNames.add(candidate);
-	return candidate;
-}
-
-/**
- * Tags page-owned content scripts in one dependency batch with a shared grouped-build id.
- *
- * @remarks
- * Lazy hydration entries (`excludeFromHtml`), component file scripts, and integration runtime
- * assets stay ungrouped. Integrations that already assign a grouped id (for example React
- * Router pages) are left unchanged.
- */
-export function assignPageOwnedContentScriptGroupedBundles(
-	dependencies: AssetDefinition[],
-	integrationName: string,
-): void {
-	if (!shouldAssignPageOwnedContentScriptGroups(integrationName)) {
-		return;
-	}
-
-	const bundleId = createPageOwnedContentScriptBundleId(integrationName);
-	const usedEntryNames = new Set<string>();
-
-	for (const dependency of dependencies) {
-		if (!isPageOwnedGroupableContentScript(dependency)) {
-			continue;
-		}
-
-		dependency.groupedBundle = {
-			id: bundleId,
-			entryName: createPageOwnedGroupedEntryName(dependency, usedEntryNames),
-		};
-	}
-}
+import type { AssetDefinition, ProcessedAsset } from './assets.types.ts';
+import { finalizeProcessedAsset } from './finalize-processed-asset.ts';
 
 /** Forces grouped content scripts to run through the bundler even in development. */
 export function ensureGroupedContentScriptsBundle(dependencies: AssetDefinition[]): void {
@@ -94,30 +12,6 @@ export function ensureGroupedContentScriptsBundle(dependencies: AssetDefinition[
 			dependency.bundle = true;
 		}
 	}
-}
-
-const KNOWN_GROUPING_INTEGRATIONS = new Set(['react', 'ecopages-jsx', 'lit', 'mdx', 'kitajs']);
-
-/** Resolves the integration name used for page-owned grouped content-script assignment. */
-export function resolveGroupingIntegrationName(processingKey: string): string | undefined {
-	if (KNOWN_GROUPING_INTEGRATIONS.has(processingKey)) {
-		return processingKey;
-	}
-
-	const prefix = processingKey.split(':')[0];
-	if (prefix && KNOWN_GROUPING_INTEGRATIONS.has(prefix)) {
-		return prefix;
-	}
-
-	if (processingKey.startsWith('ecopages-react-')) {
-		return 'react';
-	}
-
-	if (processingKey.startsWith('ecopages-jsx-')) {
-		return 'ecopages-jsx';
-	}
-
-	return undefined;
 }
 
 /**
@@ -154,7 +48,6 @@ type GroupedBundleProcessor = {
 
 type ProcessGroupedDependencyBundlesOptions = {
 	bundles: AssetDefinition[][];
-	key: string;
 	getCachedAsset: (dep: AssetDefinition, depKey: string) => ProcessedAsset | null;
 	getDependencyKey: (dep: AssetDefinition) => string;
 	getGroupedProcessor: () => GroupedBundleProcessor | undefined;
@@ -174,7 +67,6 @@ export async function processGroupedDependencyBundles(
 ): Promise<ProcessedAsset[]> {
 	const {
 		bundles,
-		key,
 		getCachedAsset,
 		getDependencyKey,
 		getGroupedProcessor,
@@ -186,7 +78,7 @@ export async function processGroupedDependencyBundles(
 	const groupedPromises = bundles.map(async (bundleDeps) => {
 		const cachedResults = bundleDeps.map((dep) => {
 			const cached = getCachedAsset(dep, getDependencyKey(dep));
-			return cached ? ({ key, ...cached } as ProcessedAsset) : null;
+			return cached ? finalizeProcessedAsset(cached, resolveProcessedAssetSrcUrl) : null;
 		});
 
 		if (cachedResults.every((result) => result !== null)) {
@@ -204,15 +96,10 @@ export async function processGroupedDependencyBundles(
 			return processedResults.map((processed, index) => {
 				const dep = bundleDeps[index]!;
 				const depKey = getDependencyKey(dep);
-				const srcUrl = resolveProcessedAssetSrcUrl(processed);
-				const processedWithKey = {
-					key,
-					...processed,
-					srcUrl,
-				};
+				const finalized = finalizeProcessedAsset(processed, resolveProcessedAssetSrcUrl);
 
-				setCachedAsset(dep, depKey, processedWithKey);
-				return processedWithKey as ProcessedAsset;
+				setCachedAsset(dep, depKey, finalized);
+				return finalized;
 			});
 		} catch (error) {
 			logError(error);
